refactor: split kunz.py into kunz/ subpackage (_coords, _vector, _fourier, _distance, _polyhedron)
